[IMPROVE] add extends Auditable

+++ src/main/java/com/enigma/entities/TransactionDetail.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @NoArgsConstructor
@@ -18,7 +19,7 @@ public class TransactionDetail extends Auditable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private Integer weight;
-    private Integer subtotal;
+    private BigDecimal subtotal;
 
     @ManyToOne(cascade = CascadeType.DETACH)
     @JsonIgnore
